dp_intent_catcher/src/requirements.txt
sentry_sdk==0.12.2
Flask==1.1.1
requests==2.22.0
nltk==3.4.5
tensorflow_hub==0.6.0
scikit_learn==0.21.3
typing==3.7.4.1
gunicorn==19.9.0
xeger==0.3.5
tqdm==4.36.1
pandas==0.25.1
openpyxl==3.0.0
